feat: implement user authentication domain entities, repository abstraction, use cases, and API routes

+++ Features/Auth/Domain/UseCases/DeleteUserUseCase.py
@@ -1,9 +1,7 @@
-from typing import Optional
+from Features.Auth.Domain.Entities.UserEntity import Role
 import logging
-from Core.Strings.RoleString import admin
 from Core.errors.AuthErrors import AuthError, UserNotHaveRole
 from Core.errors.GlobleErrors import ServerError
-from Core.security.Jwt import verify_token
 from Features.Auth.Domain.Repository.AuthRepository import AuthRepository
 
 logger = logging.getLogger(__name__)
@@ -13,14 +11,15 @@ class DeleteUserUseCase:
     def __init__(self, auth_repository: AuthRepository):
         self.auth_repository = auth_repository
 
-    async def execute(self, user_id: str, token: Optional[str] = None) -> bool:
+    async def execute(self, user_id: str, role:Role) -> bool:
         try:
-            if token is not None:
-                payload = verify_token(token)
-                role_val = getattr(payload, "role", None)
-                if role_val != admin:
-                    raise UserNotHaveRole()
-            return await self.auth_repository.delete_user(user_id)
+
+            if role == Role.USER :
+                raise UserNotHaveRole()
+
+            if role == Role.ADMIN or role == Role.SUPERADMIN :
+                return await self.auth_repository.delete_user(user_id)
+            
         except AuthError:
             raise
         except Exception as e:
